+++ modules/DateTime.py
@@ -1,5 +1,4 @@
 """Makes a table that plots gps timestamp"""
-from exif import Image
 import plotly.graph_objects as go
 
 
@@ -9,27 +8,29 @@ def date_time(photos, log):
     datetime_original = []
     datetime_digitized = []
     types = [datetime, datetime_original, datetime_digitized]
-    types_str = ["datetime", "datetime_original", "datetime_digitized"]
+    types_str = ["Creation date", "Date-time original", "Date-time digitized"]
+
+    simple_photos = []
+    for i, _ in enumerate(photos):
+        simple_photos.append(photos[i]["item"])
 
     for each in photos:
-        with open(each, 'rb') as image_file:
-            my_image = Image(image_file)
-            for i, _ in enumerate(types):
-                try:
-                    types[i].append(getattr(my_image, types_str[i]))
-                    log.debug("%s has %s data", each, types_str[i])
-                except AttributeError:
-                    log.debug("%s has no %s data ", each, types_str[i])
+        for i, _ in enumerate(types):
+            try:
+                types[i].append(each[types_str[i]])
+                log.debug("%s has %s data", each["item"], types_str[i])
+            except KeyError:
+                log.debug("%s has no %s data ", each["item"], types_str[i])
 
     fig = go.Figure(
         data=[go.Table(
             header=dict(values=[
                 "Photos",
-                "Time Stamp",
+                "Creation date",
                 "Date time Original",
                 "Date time Digitized"
                 ]),
-            cells=dict(values=[photos, datetime, datetime_original,
+            cells=dict(values=[simple_photos, datetime, datetime_original,
                                datetime_digitized]))]
         )